Changed interface I_Deduper to prevent optimisation details leaking through.
This commit changes the I_Deduper interface to return a copy of the
cells, rather than a reference to them. This gives the implementer of
I_Deduper more flexibility to how those cells are stored. For example,
the text and attributes might be separated and RLE compressed.

Buffer::getLine() was introduced as a harmonised way of getting
historical or active lines. This simplified Buffer::dispatchFg(),
Buffer::dispatchBg() and Buffer::ParaIter.